+++ src/main/java/bowling/domain/state/ScoreType.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.state;
 
 public enum ScoreType {
     STRIKE,
